Type annotation location includes colon
Summary:
The source location that Hermes currently stores for a type annotation only contains the type itself, and not the colon preceding the type annotation. This differs from the source locations for a `TypeAnnotation` node reported by the Babel parser and Flow parser ([AST Explorer](https://astexplorer.net/#/gist/66ea46f2146b7913aa4dcecf27fc78f4/634c373453713b40dc20744baac482da7f26e205)):

```
Hermes:
const x: int = 1
loc ---> ^^^

Babel, Flow:
const x: int = 1
loc -> ^^^^^
```

Let's adjust the source location for TypeAnnotation nodes in the Hermes parser to include the leading colon, for consistency with what Flow (and Babel) consider the location of a `TypeAnnotationNode`.

This involved modifying the `parseTypeAnnotation` function to take in an optional start location instead of a boolean `wrapped` argument, which if true was wrapping the parsed type annotation in a `TypeAnnotation` node with the same location as the inner type node. Then every callsite that was setting `wrapped` to true when calling `parseTypeAnnotation` now passes in the start location containing the colon. Since the start location argument is optional, if not set the parsed type is not wrapped in a `TypeAnnotation` node.

Reviewed By: avp

Differential Revision: D23881909

fbshipit-source-id: 56cbe52e112b75b861f042656f6ca237afa67f9f