Make all VTables const
Summary:
There are a few good reasons to make VTables, which are already
logically const, physically const:
* Compiler enforces constness
* Security: often put in a read-only section of memory that can't be tampered with
* Performance: compiler knows the contents of the table won't change
(I don't expect a lot from this anywhere important)

In particular, this prevents a wild write from somewhere in the program from
modifying any part of a VTable. We'll know if an invalid CellKind is in a VTable,
it is a bad VTable pointer, not some modification of the VTable.

Reviewed By: tmikov

Differential Revision: D24553780

fbshipit-source-id: 45cd5ce64666a80ef53aabc1f73fc03f299e8bed