Update esprima_test_runner to run JSX tests
Summary:
Let's update the esprima test runner to also run the esprima JSX tests.

There are only three types of currently failing tests left on the skip list:
1. Tests with HTML entities, which are implemented and removed from the skiplist in the next diff.
2. `esprima/test_fixtures/JSX/attribute-element.js` which tests JSX elements within attributes (e.g. `<a foo=<b /> />`), a feature not supported by Hermes or Flow.
3. `esprima/test_fixtures/JSX/multiline-crlf-text.js` which removes the `\n` from CLRF line endings in multiline JSX text's values. This does not appear to be standard behavior, as neither Babel, Flow, nor Hermes matches this behavior instead opting to keep the full `\r\n` or `\n`.

We also need to add the `-parse-jsx` flag if we are parsing an esprima JSX test. We can check for this easily by simply enabling the flag when the test filename contains `JSX`.

Reviewed By: avp

Differential Revision: D24464911

fbshipit-source-id: 15dc1d990da57c0f0f6242313a4fc2d12a1fd4d9