Use GC::makeA in JSObject
Summary:
Use new GC::makeA API. For JSObject, I've created a separate
constructor taking handles instead of modifying the original
constructor like I did with the other classes. I did this because the
constructor using pointers is invoked from a lot of places and is
explicitly annotated as a lightweight constructor that does not do
any allocations. Keeping a separate constructor using handles seems
like the right thing to do here but I'm open to changing it.

Reviewed By: dulinriley

Differential Revision: D24527668

fbshipit-source-id: cecbd003c6025c4fd447aa73646bcb239420745b