Respect compileFlags.debug even if debugger is not compiled in
Summary:
Currently, we ignore compileFlags.debug if the debugger is compiled out.

This is a problem for `emhermesc`, which sets the flag and serializes
the bundle. If debugger support is not built in, the resulting bundle
does not contain complete debug info. When loaded in a different
Hermes instance that *does* have debugger built in, it can't inspect
local variables due to missing lexical debug data.

Reviewed By: dulinriley

Differential Revision: D24524133

fbshipit-source-id: bf6320740881b4922ea0116951247582a997bde8