Add SATB write barrier for SymbolID
Summary:
`findYoungGenSymbols` was a hacky workaround to a real problem: if a symbol's
value is changed during marking, it is possible that Hades will miss it, and
reclaim the symbol's associated string. The same problem existed with pointers,
and the solution to that is a snapshot write barrier. We'll do the same for symbols.

Introduce a new class, `GCSymbolID`, which should be used for any symbol that lives
on the heap. It will allow assignment with a write barrier, and construction requires no
barriers. By updating the Metadata, we also ensure this is enforced by the type system.
In some cases, we know the previous value of the GCSymbolID is either invalid
(empty or deleted), or didn't exist at all (initialization in unused memory). In both these cases,
the write barrier should be skipped. In the latter case, it is required to avoid reading uninitialized
memory.

The snapshot write barrier assumes the old value was alive at the start of the collection,
just like pointers. This means we no longer need to iterate over the YG looking for symbols.

Reviewed By: avp

Differential Revision: D24572173

fbshipit-source-id: 120b9fcb2e4277487a5cd714f799b758cc6d064b