Fix missing source location for ClassProperty nodes without terminating semicolon
Summary: Currently the Hermes parser does not set an end location for `ClassProperty` nodes that are not terminated by a semicolon. Instead, we must be sure to set the end location of the `ClassProperty` node to the end location of the key, value, or type annotation depending on whether a value and/or type annotation is supplied.

Reviewed By: avp

Differential Revision: D24711040

fbshipit-source-id: bfc1de489195f8466cab96715a3f43fb7cfe5402