Fix missing source location for DeclareFunction nodes without terminating semicolon
Summary:
Currently the Hermes parser does not set an end location for `DeclareFunction` nodes that are not terminated by a semicolon. Instead, we must be sure to set the end location of the `DeclareFunction` node to the end location of the type annotation, predicate, or ending semicolon depending on whether a predicate or semicolon was supplied.

In addition, the location for the inner `Identifier`, `TypeAnnotation`, and `FunctionTypeAnnotation` nodes should end at the end of the annotation instead of the end of the entire `DeclareFunction` node. This makes sure to not include the closing semicolon (or predicate) in the `Identifier` or `TypeAnnotation`'s location.

Reviewed By: avp

Differential Revision: D24711048

fbshipit-source-id: 8a714c5182c98f655ac96c3def4952bc582aa78f