cache the last found buffer and use a std::map
Summary:
Most buffer searches will be in the same file, so it makes sense to save
the last found buffer and check for it first in the next search.

Additionally, replace the sorted array with a std::map. While it is
slower in absolute terms, it avoids any performance cliffs and the most
common case is served by the check for the last buffer anyway.

Reviewed By: avp

Differential Revision: D24678001

fbshipit-source-id: bbb066036e44dcd5c4d595847ab17b38fe8907b9