Fix -use-trace-config for synth benchmarks
Summary:
`getGCConfig` expected a the JSONObject for `runtimeConfig`, but we were
accidentally passing the root. This error was disguised in the test buy
including the GCConfig into the RuntimeConfig builder, but that doesn't
properly expose the builder interface for GCConfig to allow updates.

Remove the call to `getGCConfig` from within `RuntimeConfig`. It's now
the caller's responsibility to merge these two how they'd like.

Reviewed By: neildhar

Differential Revision: D24697015

fbshipit-source-id: 974b66bb9186055ea959bcf2a0bac8567ce91bc8