Add format option to createBCProviderFromSrc
Summary:
The current implementation of createBCProviderFromSrc is written with
the implicit assumption that the resultant bytecode will only ever be
used for direct execution, and never for serialisation. Because of
this, it contains a step where it serialises jump tables into the
opcode list so that they can be executed.

However, there are cases where we use the output of this to generate
bytecode bundles (as in compileJS). Before this diff, this had two
issues:
1. Since the jump tables are encoded inline with the opcodes, we
would end up serialising the jump tables into the bytecode twice.
Once with the opcodes, and once again when we perform the jump table
serialisation step.
2. The switch case is going to use the jump table that was encoded by
`createBCProviderFromSrc`. However, that jump table is not guaranteed
to have the 4-byte alignment we expect in memory. This is because the
padding for that jump table was computed only with respect to a
single function and not to the entire program's bytecode. As a
result, the serialised bytecode will not have the jump table at a
4-byte aligned address, and the runtime will crash when trying to
read from it.

This diff fixes the issue by allowing the caller to pass in a
`format` option to `createBCProviderFromSrc`. The format option is
used to determine whether the bytecode is intended for immediate
execution and therefore whether the jump tables should be inlined.

The fix isn't entirely ideal since it essentially involves special
casing the execution case. Instead, I've tried to address the issue
described without disturbing other callsites. A better fix would
probably involve standardising how the BytecodeModule is expected to
store data, I'm working on figuring out the right way to do that.

Reviewed By: tmikov

Differential Revision: D24075278

fbshipit-source-id: 855504abf10a95c45a13198c898ad6379ffb4021