Require call to bytecodeGenerationComplete at the end of creating a function
Summary:
Add a requirement that `bytecodeGenerationComplete` must be called
exactly once at the end of creating any function. Add asserts to
enforce this requirement. This could be done in a simpler way if we
remove all external calls to `bytecodeGenerationComplete` and instead
move the only call into `generateBytecodeFunction`. I did it this way
to minimise disturbance to existing call sites, but I'm open to other
ideas.

Now that we have a guarantee on when it will be called, we can do
more useful work in `bytecodeGenerationComplete`.

Reviewed By: tmikov

Differential Revision: D24425865

fbshipit-source-id: a4d506c5c66a329728df0eb83064b9952a999dc3