Parse HTML entities in JSX
Summary:
The Hermes parser currently does not have any special handling of HTML entities in JSX. HTML entities appear in three possible forms, each of which encodes a particular unicode code point: `&NAME;`, `&#NUMBER;`, and `&#xHEX;`. HTML entities should be converted to their equivalent unicode code points for the non-raw value of JSX strings.

This diff introduces `JSLexer::consumeHTMLEntity` which optionally returns the code point represented by either of the three entity forms described above. `JSLexer::consumeHTMLEntity` is called when the current character is an `&` within JSX text (called from `JSLexer::advanceInJSXChild`) or within a JSX string attribute (called from a JSX branch of `JSLexer::scanString`). If an HTML entity with a number (possibly in hex) is encountered, we loop through it and build up the code point. If an HTML entity with a name is encountered we instead loop through to build up the entity name, and then look it up in a map.

If `JSLexer` fails to determine a valid unicode code point for the HTML entity, the current char pointer is reset to the `&` and the caller treats the resulting bytes as if they are not an HTML entity. We make sure to handle a few edge cases, notably `&#;`, `&#x;`, and numbers that are too large to be unicode code points.

The map of entity names is initialized when constructing the `JSLexer`, using the new `HTMLEntities.def` which defines all HTML entities along with their code points using new `HTML_ENTITY` macros (this is the same list of HTML entites used by the Flow and Babel parsers).

We also make sure that the JSX entity map is only included when the `HERMES_PARSE_JSX` macro is enabled. This required adding a few more `HERMES_PARSE_JSX` guards to gate JSX-specific methods or sections of code.

Reviewed By: tmikov

Differential Revision: D24464939

fbshipit-source-id: 5aa8cef40f13df40ec062be37e006ac1d6c46daf