Add OldGen::clearFreelistForSegment
Summary:
Add a method to quickly clear the freelist for a given segment. This
is used to clear the freelist for a compactee, so that we don't add
additional objects that are going to be evacuated soon. Once
compaction is complete, the entire segment will be added back to the
freelist as a single cell.

Reviewed By: dulinriley

Differential Revision: D24386072

fbshipit-source-id: 8aabaf06aa06f8691cbc2b6582af2494e8e8f576