Fix REPL printing for adopted Promise
Summary:
This diff fixes how REPL prints "adopted" Promise. An adopted promise
is when its resolution is entirely delegated to another Promise.

Prior to this diff it would be printed as `Promise <adopted>`, which
is not very useful and the "adopted" is merely an internal
implementation detail.

To be closer to the spec and other JS REPLs, we should recurively
prettyPrint the adoptee Promise instead.

Reviewed By: avp

Differential Revision: D24804647

fbshipit-source-id: 9ca89c467ccabc0dc25905a259e85552d1efa46b