search for Python if not configured
Summary:
Search for Python if it is not configured and error out if it cannot be
found or has an unsuitable version.

Previously we simply relied in the undocumented PYTHON_EXECUTABLE flag
to be correctly set by configure.py. When possibly, the CMake build
files should not depend on other scripts and should be as self-contained
as possible.

It is usually not necessary to use configure.py. The goal of the latter
isn't to perform actual configuration (it is quite unlike the
"traditional" configure), but to document and simplify choosing
different modes. As such, it should not be required.

Reviewed By: neildhar

Differential Revision: D24584665

fbshipit-source-id: 809e5c60006b6ac1ae3e4cbc265c7f423ef6bc0c