Only update Hades's average YG survival ratio for non-promotion collections
Summary:
This change tries to improve a couple different calculations that go into
determining the "drain rate" for incremental collections:
* Don't update the YG survival ratio if nothing is collected.
This avoids skewing the distribution towards 100% when
that isn't the true survival ratio
* Initialize the OG survival ratio to 30% instead of 70% if promotion is on.
This assumes that the memory reaching OG will have the average survival ratio of a YG collection

This will prevent a worst-case scenario of a small init heap size
leading to a long blocking OG collection as soon as TTI is reached.

NOTE: These average survival ratios are only used in incremental mode,
concurrent mode ignores them. This will not affect the ratios that are
reported via GCAnalyticsEvent.

Reviewed By: neildhar

Differential Revision: D24875836

fbshipit-source-id: c97806ffef1155692e932e26fc8324959fb0ecf5