Semantic validator mode to only validate the AST without preparing for compilation
Summary:
The Semantic Validator currently performs validation of the AST, as well as transforms/preparation for further compilation. However there is a use case for running the Semantic Validator for the sole purpose of validating that the AST has a proper structure, without running validation or transforms that are specific to preparing the AST for compilation. For example we would like to be able to run these validation checks in the WASM parser as some important errors are deferred to the Semantic Validator, but do not want to run additional preparation for compilation.

Let's introduce a boolean `compile_` mode in the Semantic Validator that defaults to true. Then we simply can gate aspects of the Semantic Validator that are specific to compilation behind this flag. Only the following parts of the Semantic Validator are gated, as they are the only portions that appear to be specific to preparation for compilation:

- Transforming arrow function bodies
- Transforming try statements
- Trying to compile RegExps and erroring if they do not compile
- Erroring on private property/name nodes. These are still a valid AST structure, just aren't yet allowed for compilation.
- Transforming default exported functions without names.

All current users of the Semantic Validator will work as before, as `compile_` is set to true. All future features that we can parse but not compile will also need to be gated behind this flag.

Also introduces the `validateASTForParser` function which performs these semantic validation checks without preparation for compilation. This is currently unused but will soon be used by the WASM parser.

Reviewed By: tmikov

Differential Revision: D24903195

fbshipit-source-id: cc23f7046ce23b020da8c072fdfc1b0255311f15