Error when SemanticValidator encounters CoverTypedIdentifier
Summary:
Like the other Cover nodes, we error in SemanticValidator instead
of the parser in this case.

If we need to find a way to error in the parser these early errors in the future,
we'll have to find a way to move this check to the parser with the
other Cover validation.

Reviewed By: tmikov

Differential Revision: D23892958

fbshipit-source-id: df4f20c322c3756db7e07f53191385df66c5b7e9