Run semantic validator after parsing
Summary:
We want to run the semantic validator in the WASM parser to catch a variety of errors that are not caught in the parser itself. We can run these after the AST has been parsed and transferred to JS, and simply return an error if the semantic validator finds any errors durings its pass over the AST.

Additionally, the checks we are running with the semantic validator no longer will error on private properties (as of D24903195). We do not want to support these in the WASM parser yet though, so let's error if we encounter a `PrivateName` or `ClassPrivateProperty` in the AST adapters.

Reviewed By: gkz

Differential Revision: D24903217

fbshipit-source-id: c0e4bb602a65508af28c19c93578a42477525be8