Stop madvise MADV_DONTNEED identifier hashes (#271)
Summary:
While loading a bytecode file, Hermes will populate the identifier table
with the precomputed hashes. It then issues an madvise MADV_DONTNEED call
as it will not need to read the hashes again.

This is fine and dandy if the bytecode file is mmap'd from disk. But if it
is from memory, MADV_DONTNEED has destructive semantics: the hashes are
*zeroed*. If you then run the same bytecode file again, all identifiers
will have zero hashes, and they will be treated as distinct from the first
time the bytecode file was run.

I propose just removing the madvise call. If that is too aggressive, it
could instead be controlled by a RuntimeModuleFlag.
Pull Request resolved: https://github.com/facebook/hermes/pull/271

Test Plan: a nice new test

Reviewed By: kodafb

Differential Revision: D22029168

Pulled By: tmikov

fbshipit-source-id: cdca36bc167b7e0f5b0e33c1cff725d8119fc36a