Generate lazy data for generator functions.
Summary:
Generator functions need to be aware of laziness in the same way
other functions do. If they are not, then the inner function fails
to resolve the closure name when materializing the scope chain,
because it was never in the name table (the outer function has no
source).

Modify the laziness structure to emit data at the `genGeneratorFunction`
level instead of the inner function level,
because this prevents a situation in which the outer function is eager
(and never lazy) but contains a lexical child function which is lazy.

This requires changing the lazy data to store `isGenerator` instead
of `isGeneratorInnerFunction` and branching earlier in that pipeline.

Reviewed By: willholen

Differential Revision: D24514819

fbshipit-source-id: f8a0c52d9a42b8e4f67ef22d9694e935d01aa6e8