Detect @flow pragma to enable parsing Flow syntax
Summary:
By default Babel will detect whether an `flow` pragma is present in a source file and modify it's parsing behavior based on the presence of the Flow pragma. If the Flow pragma is not present, only syntax that is unambiguously Flow type syntax will be parsed as Flow types, whereas ambiguous syntax (e.g.`foo<T>(x)` which may be a call expression with Flow type arguments or two comparisons) will be parsed as standard JavaScript. Babel also provides a `flow: all` mode that parses all syntax as Flow type syntax without checking for the presence of a Flow pragma.

To replace uses of the Babel parser we must be able to replicate this behavior. Let's introduce a `flow` option to the WASM parser that when set to the default of `detect` will only parse ambiguous Flow syntax when a Flow pragma is present. The `flow` option can also be set to `all` in which case all potential Flow syntax is parsed as Flow types.

We must pass this option from the JS wrapper in the parser itself, and then try to detect the Flow pragma. The Flow pragma is the text `flow` in any comment that appears before the first statement (but can appear after directives). We can detect this by using the lexer to find all comments that appear before the first non-directive token (where directive tokens can be optionally followed by a semicolon). If any of these comments contains `flow` followed by a word boundary, then we consider that the Flow pragma.

If the parser is set to detect mode and Flow pragma is present in a file then we then should parse umabiguous Flow type syntax with `ParseFlowSetting::UNAMBIGUOUS` mode, otherwise we parse all Flow syntax with `ParseFlowSetting::ALL` mode.

Reviewed By: avp

Differential Revision: D24925916

fbshipit-source-id: e516c7d7ebd19dd250c68ba0a1afbafe2d22465d