Fix bugs parsing object type properties with static or proto modifiers
Summary:
There are a few bugs in the current parsing of object type annotation properties, surrounding the use of the `static` and `proto` modifiers.

## Static
There are two main issues surrounding the use of the `static` modifier. First, we should allow a variance sigil after the `static` keyword like so (currently the parser errors in this case):
```
declare class C {
  static +foo: string
}
```

We are also not erroring when the static keyword is used outside `declare class` statements. All of the following object type properties should error due to an invalid static modifier:
```
interface I {
  static foo: string       // Properties
  static foo(): void       // Method types
  static (): void          // Callables
  static [string]: string  // Indexers
  static [[foo]]: string   // Internal slots
  static get foo(): string // Getters
}
```

## Proto
First of all, the `proto` modifier is only allowed within `declare class` statements, and should not be allowed in interfaces. This means that using the `proto` modifier in an inteface should error due to an invalid proto modifier, like in the following case:
```
interface I {
  proto foo: string
}
```

Due to the above we are also parsing the following incorrectly, as it is currently being parsed as a callable property with the `proto` modifier instead of the correct method type with name `proto`:
```
interface T {
  proto(): void
}
```

Finally, the `proto` modifier is only allowed on regular properties, and should error on all of the following cases. (Currently `proto` correctly errors on only indexers and method type properties):
```
declare class C {
  proto (): void          // Callables
  proto [[foo]]: string   // Internal slot
  proto get foo(): string // Getters
  proto+: string          // Malformed variance
}
```

Reviewed By: avp

Differential Revision: D24967521

fbshipit-source-id: 0205ec8e805ca1617c8772b44fa2ae89b99bd5b5