Add emitStep
Summary:
This diff adds a new function for regex codegen called emitStep.
emitStep should be called repeatedly and each time it will return a
`NodeList*` corresponding to a set of nodes that must be emitted
before the next call to emitStep. Once emitStep returns a `nullptr`,
you can stop calling it.

This diff also modifies the base version of `emit` to start using
emitStep and modifies some simple `emit` functions to use the new
interface.

Reviewed By: avp

Differential Revision: D22127418

fbshipit-source-id: 994e5b25f951523b8aa74db822e373582281e26f