Generate separate IR functions for copies of the same CommonJS module
Summary:
Generates separate IR functions for copies of the same CommonJS module in multiple segments. By definition, all copies of a CommonJS module are interchangeable at runtime. However, sharing the *IR* across these copies results in incorrect bytecode because of some apparent bad interaction with lowering / register allocation. So now we don't do that anymore.

As a side effect, this diff eliminates the `Context::SegmentInfo` struct (fka `SegmentRange`) in favour of storing just the segment IDs in `Context`. The mapping of modules to segments is kept entirely in the IR.

Reviewed By: avp

Differential Revision: D25024634

fbshipit-source-id: 3fbf8e9e61aff6aab8456eea4d85a7c259f1ef01