Remove heap size test from android instrumentation tests
Summary:
This test doesn't work with Hades, which doesn't set it's "Heap size" to be
the "target capacity" of the heap, but instead its current used bytes.

This test relies on it reporting a different number for a different initial heap size,
which Hades doesn't do. Delete the test, because changing Hades's reported stats
here would be misleading.

Reviewed By: neildhar

Differential Revision: D25257368

fbshipit-source-id: 97b33e18f1000fa01aafd585c82bc21fca4b6258