Add accept overloads for GCHermesValue and PinnedHermesValue
Summary:
Distinguish between `GCHermesValue` and `PinnedHermesValue` in
`SlotAcceptor`. This lets us write acceptors that are aware of
whether a given pointer is in the heap. This can then be used by
compaction when figuring out whether to dirty a card.

Add an `acceptHV` function to `SlotAcceptorDefault` that can be
used the same way our current `HermesValue` overload works. But
it uses a different name to ensure it doesn't get called accidentally.

Reviewed By: dulinriley

Differential Revision: D25265423

fbshipit-source-id: 6f21a6945796d3f2b1ea25008dab4c04240b279b