Support -Werror=<category>
Summary:
Hermes currently supports `-Werror` (and `-Wno-error`) which causes all warnings (or no warnings) to be treated as errors. This diff adds support for the clang-inspired flags `-Werror=<category>` (and `-Wno-error=<category>`) to do this for each individual warning category.

This diff also centralises the warning category definitions in a `Warnings.def` file and uses macro expansion to generate:
* The `hermes::Warning` enum.
* `-W<category>` and `-Wno-<category>` flags, and corresponding `setWarningStatus()` calls, for all categories that aren't hidden.
* `-Werror=<category>` and `-Wno-error=<category>` support for all categories.

Reviewed By: tmikov

Differential Revision: D24555509

fbshipit-source-id: 78a44046d4ccfc9c4ca32c164e3c8f172113d0cf