don't release the stack trace
Summary:
Setting the "stack" property from JS was clearing the `stacktrace_` C++
field. In certain circumstances that would cause a null pointer
dereference in native code and a crash.

The following JS code illustrates the problem:

    var errorObj = new Error();
    Object.defineProperty(errorObj, "message", {
      get: function () { errorObj.stack = "foobar"; },
    });
    print(errorObj.stack);

The string representation of the "stack" property is constructed lazily.
It is initially a getter which on its first invocation constructs the
string and reconfigures itself as a normal property (not a great design,
it would have been better to stay as a getter but just return the
string; fixing that however is beyond the scope of this diff).

Constructing of the string starts by checking whether "stacktrace_" is
null, and returning an empty string in that case. It then continues to
obtaine the message property, which, as we can see above, can maliciously
be configured as a getter modifying the "stack" property, which in turn
clears the "stacktrace_" field. So, we end up with "stacktrace_" being
unexpectedly cleared "under us", after we have already checked it.

The simplest fix is simply to never clear the `stacktrace_` field once
set. It is actually a nice invariant and that is what we went with.

Other fixes were considered, but not found to be worthwhile:
- Keeping a reference to "stacktrace_" when we check it. That would
  involve changing it to a std::shared_ptr.
- Reordering the accesses so that we obtain the "message" property after
  we are done with "stacktrace_". That would require making more string
  copies.

Reviewed By: avp

Differential Revision: D25320589

fbshipit-source-id: 8fc490a0113e4bdd03c98d608a0f2b041cf94f0f