Add detachedness to heap snapshots
Summary:
Chrome has added a new "detachedness" field to nodes in v87 and
above. Hermes should output this additional field.
In Chrome, "detachedness" of 1 means that an object is not attached to
a DOM tree. This is irrelevant to Hermes, so we'll always output 0,
which means every node is attached.

Reviewed By: neildhar

Differential Revision: D25290008

fbshipit-source-id: 33ccb5a1c7284342d30e5fa53594108c720c5c24