Calcuate end source location from the previous location instead of eatSemi
Summary:
The Hermes parser is fixing its handling of source locations around parenthesized types and expressions (choosing to follow the same rules as Flow and Babel). When an expression or type node is parenthesized, the source location of the node correctly does not contain the parentheses. However currently when a parent node has a parenthesized expression or type child node, Hermes is inconsistently including the child's parens in the parent's source location. Hermes should instead always include the child's parentheses in the parent node's source location. For example:

```
(1 + 2) * (3 + 4)
 ^^^^^     ^^^^^  // Left and right inner expr locs do not contain parens
^^^^^^^^^^^^^^^^^ // Parent binary expression node does contain parens of children
```

We also can now check the end location of the previous token using `getPrevTokenEndLoc`, which will be used to calculate source locations in this scheme instead of building up source locations from child nodes (as child nodes will not contain their own parentheses in their source locations).

In preparation for this let's migrate usages of the `eatSemi` helper to instead use `getPrevTokenEndLoc` for calculating the end source location. Currently `eatSemi` takes a reference to the current end source location for a node as its first argument, and sets that loc to the end of the semicolon if one was consumed. This is used to properly set the end source location after the semicolon if one exists.

However now that we have `getPrevTokenEndLoc` the end location does not need to be explicitly maintained or passed into `eatSemi`. Instead we can just make a single call to `getPrevTokenEndLoc` when building the node, which will automatically return either the end location of the semicolon if a semicolon exists, or the end location of the token before it if not.

Not all of the uses of `eatSemi` are for a node that ends with an expression or type (which would necessitate the use of `getPrevTokenEndLoc` for proper paren locs). However it is simpler if we just change the API for every use of `eatSemi` to not take in the end source location, and even cases where `getPrevTokenEndLoc` is not strictly needed yet are still often simplified due to not needing to maintain the end location to be passed into `eatSemi`.

Reviewed By: tmikov

Differential Revision: D25291370

fbshipit-source-id: 9af351229ac01e42c9989ec23fe1d0fa23ebd11a