AST nodes with parenthesized expression child nodes include child's parens in source location
Summary:
The Hermes parser is fixing its handling of source locations around parenthesized types and expressions (choosing to follow the same rules as Flow and Babel). When an expression or type node is parenthesized, the source location of the node correctly does not contain the parentheses. However currently when a parent node has a parenthesized expression or type child node, Hermes is inconsistently including the child's parens in the parent's source location. Hermes should instead always include the child's parentheses in the parent node's source location. For example:

```
(1 + 2) * (3 + 4)
 ^^^^^     ^^^^^  // Left and right inner expr locs do not contain parens
^^^^^^^^^^^^^^^^^ // Parent binary expression node does contain parens of children
```

We also can now check the end location of the previous token using `getPrevTokenEndLoc`, which will be used to calculate source locations in this scheme instead of building up source locations from child nodes (as child nodes will not contain their own parentheses in their source locations).

This diff fixes parenthesized source locations for all remaining AST nodes in `JSParserImpl.cpp`. The rules are as follows:

1. If a node can begin with a type or expression node in the source text, then we must save the start source location before parsing that initial type or expression. This will be the start location of the first paren that wraps that type/expression node, if any exist.
2. If a node can end with a type or expression node in the source text, then we must call `getPrevTokenEndLoc` after the last type/expression node has been parsed. This will be the end location of the first paren that wraps that type/expression node, if any exist.

This diff then goes through and implements these rules for every remaining AST node that can begin or end with a type or expression node. For the most part this is a straightforward translation, however there are two main places that require more thought:

Support for parenthesized children in `BinaryExpression` and `LogicalExpression` nodes is a bit more involved, since the parser uses a stack of nodes to account for operator precedence. Here we need to store the start location (before parentheses) for each expression in the stack, as these expressions will be placed on the left hand side. We also need to keep track of the end location (after parentheses) for the "top expression", as it will always be placed on the right hand side. We then pass these start and end locs to the `newBinNode` function to properly set source locations.

The other place of note is for locations of new/member/call expression chains. When an expression is passed to a another parse function, we also must pass the start loc (before parentheses) for that expression. There are also a few loops in these parse functions where we must make sure to preserve the start location (and sometimes split it up with the `objectLoc` which is used for errors/as a debug loc, and is initialized to the start loc but changes within these loops).

In addition this diff replaces some cases where the end location is explicitly calculated from multiple potential end nodes, with a single call to `getPrevTokenEndLocation` at the end of the node (e.g. for import specifiers).

Reviewed By: avp

Differential Revision: D25291388

fbshipit-source-id: fdf4e7090c48db197e64eb5d4f2c00b89a900e9a