AST option to skip "raw" properties.
Summary:
The "raw" AST property is intended to show the value of a number exactly
as it was entered originally (for example in hex). While occasionally
useful, it starts getting in the way when we are transforming AST.

Introduce a new hidden option `-Xinclude-raw-ast-prop` (enabled by
default), so it can be optionally disabled.

A test is included in the next diff.

Reviewed By: avp

Differential Revision: D25295348

fbshipit-source-id: bfd1cc252c48eb2588ea4898853d7f1ec33cca2b