Fix StringPrimitive memcpy test for GCC
Summary:
We have a test for StringPrimitive that tests whether `std::string`
and `std::u16string` can be safely `memcpy`'d. However, the string
object may contain a large number of uninitialised bytes used for
storing a small string. In order to compare the value before and
after a move, the test uses a trick where it `memcpy`'s the first
string object into a buffer, before move-constructing the second
string in that buffer. This ensures that the untouched bytes remain
the same.

However, the C++ standard does not guarantee that placement new into
a buffer will leave untouched bytes alone. It appears GCC is able to
use this to optimise away the `memcpy` trick in anything above -O0,
causing the test to fail.

Here is a simple demo: https://godbolt.org/z/6nqb6f

To fix this, make a heap allocated buffer that is zeroed out with `calloc`.
Then we can allocate into that zeroed out buffer. This is still technically
not guaranteed to work, since placement new is still allowed to modify
the buffer, but it seems to be good enough for GCC.

Reviewed By: tmikov

Differential Revision: D25359225

fbshipit-source-id: 92b660c54905e71537916a04abbb728a189fec49