Allow storing tokens in the lexer.
Summary:
Add a token storage which contains `StoredToken`.
It tracks the kind and source range of all tokens,
but can be augmented to add more information if required.

The complexity comes from backtracking.
When using `SavePoint`, we save the size of the `tokenStorage_` and
restore it as well as the token.
When using `lookahead1`, we `pop_back` when we revert to the original token.

Local testing on 100+ MB JS bundles has shown no detectable regression
in total parsing time when this feature isn't enabled at runtime.

Reviewed By: tmikov

Differential Revision: D23939696

fbshipit-source-id: a9f059fb97dee9d693124b2fe73b022f619ff756