Track per-segment peak allocated bytes
Summary:
Track the peak allocated bytes for each segment. For now, this is an
approximation, since we only update the peak right before sweeping.
It is theoretically possible that the allocated byte count will grow
past the peak between sweeping and the time this value is accessed.

This peak is therefore only accurate as of the last time the segment
was swept.

Reviewed By: dulinriley

Differential Revision: D24490415

fbshipit-source-id: fc766b29de1edc74cf80d156dfc19d23704f16d6