Avoid covering typed parameters as yield arguments.
Summary:
`yield` arguments must not eat `:` after an assignment expression
in an attempt to parse typed identifiers for an arrow function expression,
because this leads to bugs in code such as:
```
x : yield {} : 1
```
when the `: 1` is eaten as part of the assignment expression which is
the argument to `yield`.

Pass the correct arguments to `parseAssignmentExpression` and fix
this problem.

Reviewed By: tmikov

Differential Revision: D24626904

fbshipit-source-id: f7ac8f6d55ceb7030f2f483cb6377027d5a3322e