Efficient serialization scheme for transferring AST from WASM to JS
Summary:
The slowest part of the WASM parser is transferring the AST from WASM to JS. Much of this performance overhead comes from the frequent calls made between WASM/JS to build up the AST in JS node by node. We can improve upon this performance significantly with a custom serialization scheme that allows us to no longer need any intermediate calls across JS/WASM. In current performance tests this new serialization scheme results in a speedup of the WASM parser by a factor of ~1.8x.

## Overview

At a high level this replaces the `HermesParserJSBuilder` in WASM with a `HermesParserJSSerializer` that produces two buffers - a buffer containing the serialized AST and a buffer containing serialized source positions. The structure of the serializer closely matches the structure of the old JS builder - we first perform a traversal over the entire AST, writing a serialized version of the AST to the "program buffer" in pre-order. During this pass we store the start and end position of each source location (like before), and then reuse the logic from before to calculate the UTF-16 JS source locations. These calculated source positions are stored in a "position buffer", and are associated with a specific node's source location by a unique loc id that we generate and store during the initial pass over the AST.

On the JS side, we no longer are making calls from WASM to JS so we no longer need to supply a JS library that is linked into WASM. Instead we use a new `HermesParserJSDeserializer` to deserialize the AST and its source locations from the aforementioned program and position buffers. Like the JS library before it, all node deserializers are generated directly from the Hermes AST definition via macros. We can deserialize the entire AST in a single pass through the program buffer to construct the AST with empty source locations, followed by a single pass through the positions buffer to fill in those calculated source locations. We simply must keep create a map from loc ids to the source location object on the first pass so that they can be filled in during the pass through the positions buffer.

## Serialization scheme

The exact serialization scheme can be seen in the code, but the following is a quick summary:

Accessing values in the WASM heap from JS requires that the values be aligned. Due to this the scheme would be greatly complicated if we were writing values to the buffer of different sizes, so we choose to write values in 4-byte chunks. This allows us to represent addresses, sizes, booleans, and node/comment types as a single 4-byte value. The only complication is for 8-bytes doubles which we can write as two 4-byte chunks as long as they are aligned on 8-byte boundaries.

- Nodes are serialized as their node kind converted to an integer, followed by their loc id, followed by their serialized properties.
- Node lists are serialized as their size followed by their serialized elements.
- Strings are serialized as their pointer followed by their size.
- Booleans are serialized as their value (converted to a 4-byte int).
- Numbers are 8 bytes, so we serialize them in two 4-byte chunks but make sure to prepend 4 bytes of padding to align them on 8-byte boundaries if necessary.

When deserializing, we use the NodeKind to index into an array of generated deserialization functions for each kind of AST node.

## Additional notes

- We cannot free the Context or Parser until deserialization is complete, so add them to the ParseResult before deserialization so that they will be freed when the entire ParseResult is freed after deserialization is complete.
- We must include `ESTREE_NODE_FIRST` and `ESTREE_NODE_LAST` entries in the array of node deserializers so that the node kind can be used as an index, since they both are included in the NodeKind enum.
- Now that we don't need a JS library or any pre-js files for emscripten, the build process can be simplified.

Reviewed By: tmikov

Differential Revision: D25548235

fbshipit-source-id: 1a3f71c6456025950f47ebb57c016fe5f02b5d6c